requirement.txt
streamlit
tensorflow
pillow
numpy
os